Do not add VPN local exclusion rules
Prevent UIDs on a VPN from accessing the private IP ranges of networks
that they are not allowed to access. Without this, when connected to a
VPN, apps that are disallowed from accessing a Wi-Fi network will
become able to access the Wi-Fi network's LAN despite not being able
to access the Wi-Fi network whatsoever before connecting to a VPN.

Stop adding the local exclusion rule that makes this bypass possible.

Change-Id: I9975b5ab1306ee86863979d1fe73203799cce648
Signed-off-by: Mohammad Hasan Keramat J <ikeramat@protonmail.com>